Update configuration object instead of just replacing the file
The configuration that the web app sends will likely not have all of the
same fields as the configuration that we store on the device; thus we
need to merge the two configurations together --- plus, it's just the
right thing to do. The deepmerge package from npm is used to merge the
configurations together.

A placeholder function for setting model data was added so the web app
could send values