Get AWS downloads working again
"Working" is a slightly loose term here. Manually posting to the
download endpoint on 127.0.0.1:3355 successfully downloads and installs
overlays now. The web app's AWS functionality is broken, though.

Manually sending json to the endpoint only works on 127.0.0.1 now
because the server is only listening on that address; this is not ideal
right now, but would not be a huge deal if the web app still worked.